+++ lib/bloc/counter/counter_bloc.dart
@@ -12,23 +12,4 @@ class CounterBloc extends Bloc<CounterEvent, CounterState> {
   }
 }
 
-class DragCounterBloc extends Bloc<DragCounterEvent, DragCounterState> {
-  DragCounterBloc() : super(DragCounterState(0, 1)) {
-    on<DragCounter>(
-      (event, emit) => emit(
-        DragCounterState(
-          state.dragOffset + event.dragOffset,
-          event.scale,
-        ),
-      ),
-    );
-    on<ResetDragCounter>(
-      (event, emit) => emit(
-        DragCounterState(
-          0,
-          1.0,
-        ),
-      ),
-    );
-  }
-}
+
