+++ lib/bloc/counter/counter_event.dart
@@ -1,0 +1,9 @@
+// Counter event...
+abstract class CounterEvent {}
+
+class Increment extends CounterEvent {}
+
+class Decrement extends CounterEvent {}
+
+class ResetCounter extends CounterEvent {}
+
