+++ lib/bloc/counter/counter_state.dart
@@ -1,0 +1,5 @@
+class CounterState {
+  final int counterValue;
+  CounterState(this.counterValue);
+}
+
