+++ lib/bloc/drag_counter/darg_counter_state.dart
@@ -1,8 +1,3 @@
-class CounterState {
-  final int counterValue;
-  CounterState(this.counterValue);
-}
-
 class DragCounterState {
   final double dragOffset;
   final double scale;
