+++ lib/bloc/drag_counter/drag_counter_event.dart
@@ -1,12 +1,3 @@
-// Counter event...
-abstract class CounterEvent {}
-
-class Increment extends CounterEvent {}
-
-class Decrement extends CounterEvent {}
-
-class ResetCounter extends CounterEvent {}
-
 // Drag Event...
 abstract class DragCounterEvent {}
