+++ lib/main.dart
@@ -1,15 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:toropal_counter/view/counter_page.dart';
-
-import 'counter_bloc.dart';
+import 'package:toropal_counter/bloc/index.dart';
+import 'package:toropal_counter/presentation/index.dart';
 
 void main() {
-  runApp(const MyApp());
+  runApp(const ToropalCounterApp());
 }
 
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
+class ToropalCounterApp extends StatelessWidget {
+  const ToropalCounterApp({super.key});
 
   @override
   Widget build(BuildContext context) {
