+++ lib/presentation/index.dart
@@ -1,0 +1,7 @@
+export 'package:toropal_counter/presentation/screen/counter_page.dart';
+
+export 'package:toropal_counter/presentation/widgets/counter_component.dart';
+export 'package:toropal_counter/presentation/widgets/counter_increment_decrement_button.dart';
+export 'package:toropal_counter/presentation/widgets/counter_text.dart';
+export 'package:toropal_counter/presentation/widgets/information_indicator.dart';
+export 'package:toropal_counter/presentation/widgets/reset_counter_button.dart';
