+++ lib/presentation/widgets/counter_component.dart
@@ -3,11 +3,8 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:toropal_counter/counter_event.dart';
-import 'package:toropal_counter/counter_state.dart';
-import 'package:toropal_counter/view/counter_increment_decrement_button.dart';
-
-import '../counter_bloc.dart';
+import 'package:toropal_counter/presentation/index.dart';
+import 'package:toropal_counter/bloc/index.dart';
 
 class CounterComponent extends StatelessWidget {
   const CounterComponent({super.key});
@@ -74,51 +71,8 @@ class CounterComponent extends StatelessWidget {
                     ],
                   ),
                 ),
-                AnimatedPositioned(
-                  left: dragBloc.state.dragOffset >= 0 ? null : maxDistance,
-                  right: dragBloc.state.dragOffset <= 0 ? null : maxDistance,
-                  duration: const Duration(milliseconds: 100),
-                  curve: Curves.decelerate,
-                  child: Container(
-                    width: 80.0,
-                    height: 80.0,
-                    decoration: BoxDecoration(
-                      color: const Color(0xFF2c2c2c),
-                      border: Border.all(
-                          width: 2.0,
-                          color: maxDistance == 10.0
-                              ? dragBloc.state.dragOffset > 0
-                                  ? Colors.green
-                                  : Colors.red
-                              : Colors.white24),
-                      shape: BoxShape.circle,
-                    ),
-                    child: AnimatedSwitcher(
-                      duration: const Duration(milliseconds: 300),
-                      transitionBuilder:
-                          (Widget child, Animation<double> animation) {
-                        return ScaleTransition(
-                          scale: animation,
-                          child: child,
-                        );
-                      },
-                      child: Transform.scale(
-                        scale: dragState.scale,
-                        child: BlocBuilder<CounterBloc, CounterState>(
-                          builder: (context, counterState) {
-                            return Text(
-                              '${counterState.counterValue}',
-                              style: const TextStyle(
-                                color: Colors.white,
-                                fontSize: 30.0,
-                                fontWeight: FontWeight.bold,
-                              ),
-                            );
-                          },
-                        ),
-                      ),
-                    ),
-                  ),
+                CounterText(
+                  dragBloc: dragBloc,
                 ),
               ],
             ),
