+++ lib/presentation/widgets/information_indicator.dart
